refactor: '//' in imports to '/'
