+++ src/apps/backend/modules/task/internal/task-util.ts
@@ -1,5 +1,5 @@
 import { Task } from 'backend/modules/task';
-import { TaskDB } from 'backend/modules/task/internal//store/task-db';
+import { TaskDB } from 'backend/modules/task/internal/store/task-db';
 
 export default class TaskUtil {
   public static convertTaskDBToTask(taskDb: TaskDB): Task {
